Show no issues found if no issues found

+++ app/src/main/java/de/taz/app/android/ui/settings/SettingsFragment.kt
@@ -311,34 +311,40 @@ class SettingsFragment : BaseViewModelFragment<SettingsViewModel>(R.layout.fragm
 
         val feedName = downloadedIssueStubList.firstOrNull()?.feedName ?: DISPLAYED_FEED
 
-        deletionProgress.visibility = View.VISIBLE
-        deletionProgress.progress = 0
-        deletionProgress.max = downloadedIssueStubList.size
-        deletionProgressText.visibility = View.VISIBLE
-        deletionProgressText.text = getString(
-            R.string.settings_delete_progress_text,
-            counter,
-            deletionProgress.max
-        )
+        if (downloadedIssueStubList.isNotEmpty()) {
+            deletionProgress.visibility = View.VISIBLE
+            deletionProgress.progress = 0
+            deletionProgress.max = downloadedIssueStubList.size
+            deletionProgressText.visibility = View.VISIBLE
+            deletionProgressText.text = getString(
+                R.string.settings_delete_progress_text,
+                counter,
+                deletionProgress.max
+            )
 
-        for (issueStub in downloadedIssueStubList) {
-            try {
-                contentService.deleteIssue(issueStub.issueKey)
-                counter++
-                deletionProgress.progress = counter
-                deletionProgressText.text = getString(
-                    R.string.settings_delete_progress_text,
-                    counter,
-                    deletionProgress.max
-                )
-            } catch (e: CacheOperationFailedException) {
-                val hint = "Error while deleting ${issueStub.issueKey}"
-                log.error(hint)
-                e.printStackTrace()
-                Sentry.captureException(e, hint)
-                toastHelper.showSomethingWentWrongToast()
-                break
+            for (issueStub in downloadedIssueStubList) {
+                try {
+                    contentService.deleteIssue(issueStub.issueKey)
+                    counter++
+                    deletionProgress.progress = counter
+                    deletionProgressText.text = getString(
+                        R.string.settings_delete_progress_text,
+                        counter,
+                        deletionProgress.max
+                    )
+                } catch (e: CacheOperationFailedException) {
+                    val hint = "Error while deleting ${issueStub.issueKey}"
+                    log.error(hint)
+                    e.printStackTrace()
+                    Sentry.captureException(e, hint)
+                    toastHelper.showSomethingWentWrongToast()
+                    break
+                }
             }
+        } else {
+            // no issues found:
+            deletionProgressText.visibility = View.VISIBLE
+            deletionProgressText.text = getString(R.string.settings_delete_no_more_issues_found)
         }
         // clean up file system:
         storageService.deleteAllUnusedIssueFolders(feedName)
